requirements.txt
# VidStega - Production Dependencies
# Video Steganography using LSB

# Core Flask
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-SocketIO>=5.3.0
Werkzeug>=2.3.0

# Async Task Queue
celery>=5.3.0
redis>=4.5.0

# Video Processing
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
moviepy>=1.0.3
numpy>=1.24.0
Pillow>=10.0.0

# Cryptography
cryptography>=41.0.0
pycryptodome>=3.18.0

# Error Correction
reedsolo>=1.7.0

# WebSocket
python-socketio>=5.8.0
eventlet>=0.33.0

# Utilities
python-dotenv>=1.0.0
gunicorn>=21.0.0

# Development (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
